+++ src/main.rs
@@ -106,6 +106,7 @@ fn main() -> Result<()> {
                 ))
             },
         }
+        println!();
     }
 
     Ok(())